Refactor shared buffers
